% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destinos_ruta.R
\name{destinos_ruta}
\alias{destinos_ruta}
\title{Devuelve los destinos de una ruta particular}
\usage{
destinos_ruta(ruta)
}
\arguments{
\item{ruta}{}
}
\value{
json
}
\description{
Devuelve los destinos de una ruta particular
}
\examples{
 destinos_ruta("13 Santa Eulària > Eivissa")

}
