% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradas.R
\name{paradas}
\alias{paradas}
\title{Devuelve las paradas del servicio de autobús de la isla de Ibiza}
\usage{
paradas()
}
\value{
json
}
\description{
Devuelve las paradas del servicio de autobús de la isla de Ibiza
}
\examples{
 paradas()

}
