% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntos_ruta.R
\name{puntos_ruta}
\alias{puntos_ruta}
\title{Devuelve los puntos de una ruta para una ruta y destino}
\usage{
puntos_ruta(ruta, destino)
}
\arguments{
\item{ruta, destino}{}
}
\value{
json
}
\description{
Devuelve los puntos de una ruta para una ruta y destino
}
\examples{
 puntos_ruta("13 Santa Eulària > Eivissa","Eivissa/CETIS")

}
