% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutas.R
\name{rutas}
\alias{rutas}
\title{Devuelve las rutas del servicio de autobús de la isla de Ibiza}
\usage{
rutas()
}
\value{
json
}
\description{
Devuelve las rutas del servicio de autobús de la isla de Ibiza
}
\examples{
 rutas()

}
