% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempo_por_parada.R
\name{tiempo_por_parada}
\alias{tiempo_por_parada}
\title{Devuelve el tiempo restante de la parada seleccionada}
\usage{
tiempo_por_parada(parada)
}
\arguments{
\item{parada}{}
}
\value{
json
}
\description{
Devuelve el tiempo restante de la parada seleccionada
}
\examples{
 tiempo_por_parada()

}
