% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempos_paradas.R
\name{tiempos_paradas}
\alias{tiempos_paradas}
\title{Devuelve el listado de las paradas de una ruta y destino con los tiempos restantes y lat, long de las paradas}
\usage{
tiempos_paradas(ruta, destino)
}
\arguments{
\item{ruta, destino}{}
}
\value{
json
}
\description{
Devuelve el listado de las paradas de una ruta y destino con los tiempos restantes
}
\examples{
 tiempos_paradas("13 Santa Eulària > Eivissa","Eivissa/CETIS")

}
